Replace Travis and AppVeyor with GitHub Actions.  (#250)
* First attempt at using github actions.

* Turn off appveyor CI for now. Add github actions workflow.

* Fix github actions workflow.

* Another github actions attempt.

* Github actions: add osx build

* Github actions: another attempt.

* Github actions: another try.

* Github actions: more build options.

* Github actions: another try.

* Github actions: another try.

* Update build.yml

* Update build.yml

* Update build.xml

* Update build.yml

* Update build.yml and build script.

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Update build.yml

* Remove python-devel from centos for now

* Update build.yml

* Update build.yml

* QFS C library: eliminate warning.

* Test scripts: bump x509 key sizes in order to make test work on centos 8.

* * Makefile: use /etc/system-release if present to extract os version on centos/redhat.
* Build.yml: add experimental docker build.
* Build script: do not build old Hadoop on Debian.

* * Fix docker build
* Use sha256 instead of sha1 to make test work on Debian 10.

* Build.yml use sudo to install curl

* * Makefile: fix tarball target by adding missing escape
* Build.yml: run all builds under docker do not install curl as it should be already installed
* Qfsmkcerts.sh: Use weak keys/md with older openssh.

* * Install diff utils on centos8 in order to get cmp
* Pass github env. variables to codecov.
* Do not use Travis specific env. variable to determine os build type.

* Build script: fix typo.
Makefile: better cenos release version parsing.

* Build script: fix syntax error.

* * Use only major version for centos tarballs.
* Add experimental tarball upload.
* Add Ubuntu 20.04 build

* Fix tarball upload condition.

* Fix ubuntu 20.04 hang by turning on non interactive install mode.

* Set env. var with bash line.

* Eliminate gcc9 unaligned warning.

* Fix %s output truncation gcc warning.

* Restrict ubuntu 20.04 to building the same Hadoop versions as 18.04.

* Build.yml upload build result to AWS bucket.

* Build.yml: add cygwin build

* Build.yml: update cygwin build.

* Build.xml: cygwin build another try.

* Build.yml: another cygwin try.

* Build.yml: debug cygwin build

* Build.yml: another cygwin build attempt

* Build.yml: cygwin build.

* Build.yml: fix cygwin build: checkout code

* Build.yml: cd into code.

* Build.yml: cygwin build: set git to use LF

* Build.xml: cygwin force -flax-vector-conversions

* Build.yml fix cygwin dependecies by removing space

* Cygwin build: add minimal test and s3 upload.

* Meta server: add instrumentation to track dumpster creation failure on cygwin.

* Meta server: more instrumentation to track down dumpster directory creation failure on Cygwin.

* More instrumentation for Cygwin fs creation failure tracking.  Do debug build on cygwin for now.

* Revert Cygwin to release build; attempt to add localhost to /etc/hosts

* Add new line at the end of /etc/hosts file on Cygwin.

* Try -O1 g++ optimization on cygwin.

* * Use debug build on Cygwin with -O1 flag.
* Remove instrumentation used to track gcc optimization bug.

* Just do debug build wiht no optimization on Cygwin as even -01 does not produce usable meta server executable.

* * Fix compile error.
* Use powershell to call aws cli.

* Fix /etc/hosts edit on Cygwn

* Another attemp to fix cygwin build.

* Use printf instead of echo to edit hosts and insert \r\n

* Fix Cygwin make optins.

* Fix conditional uploads.
Turn off test on Cygwin for now.

* Partially revert last change.

* Fix syntax error

* Build.yml: fix conditional upload.

* Build.xml: change upload condition to upload only form master and tagged branches.

* Build.xml:
* Fix if conditions
* Turn off /etc/hosts edit on Cygwin for now

* Build.xml an attempt to make conditional uploads work

* Try to get if statements right

* Try another if variant

* Anotehr multiline if

* Remove &&

* Another if

* single line if

* another try

* simpler if

* One more if variant

* another try

* check if matrix works properly

* test upload

* Fix typo

* Fix test upload

* Build.xml: re-enable all bulds.

* Wiki: update binary distributions.

* Rename back .appveyor

* Remove appveyor.yml

* Remove travis.yml

* README: update build status urls

* Appveyor: turn off build.

* Comment out appveor install

* Build.yml change permissions on .git in order to eliminate warning.

+++ travis/script.sh
@@ -30,11 +30,12 @@ DEPS_UBUNTU=$DEPS_UBUNTU' passwd curl openssl fuse gdb'
 DEPS_DEBIAN=$DEPS_UBUNTU
 
 DEPS_CENTOS='gcc-c++ make git boost-devel krb5-devel'
-DEPS_CENTOS=$DEPS_CENTOS' python-devel fuse-devel java-openjdk java-devel'
+DEPS_CENTOS=$DEPS_CENTOS' fuse-devel java-openjdk java-devel'
 DEPS_CENTOS=$DEPS_CENTOS' libuuid-devel curl unzip sudo which openssl fuse gdb'
 
 DEPS_CENTOS5=$DEPS_CENTOS' cmake28 openssl101e openssl101e-devel'
 DEPS_CENTOS=$DEPS_CENTOS' openssl-devel cmake'
+DEPS_CENTOS8=$DEPS_CENTOS' diffutils'
 
 MYMVN_URL='https://downloads.apache.org/maven/maven-3/3.6.3/binaries/apache-maven-3.6.3-bin.tar.gz'
 
@@ -158,7 +159,7 @@ init_codecov()
     # Pass travis env vars to code coverage.
     mkdir -p  "$MYTMPDIR"
     {
-        env | grep -E '^(TRAVIS|CI)' | sed \
+        env | grep -E '^(TRAVIS|CI|GITHUB)' | sed \
             -e "s/'/'\\\''/g"  \
             -e "s/=/=\'/" \
             -e 's/$/'"'/" \
@@ -184,8 +185,9 @@ install_maven()
 build_ubuntu()
 {
     $MYSUDO apt-get update
-    $MYSUDO apt-get install -y $DEPS_UBUNTU
-    if [ x"$1" = x'18.04' ]; then
+    $MYSUDO /bin/bash -c \
+        "DEBIAN_FRONTEND='noninteractive' apt-get install -y $DEPS_UBUNTU"
+    if [ x"$1" = x'18.04' -o x"$1" = x'20.04' ]; then
         QFSHADOOP_VERSIONS=$MYQFSHADOOP_VERSIONS_UBUNTU1804
     fi
     if [ x"$1" = x'14.04' ]; then
@@ -204,6 +206,11 @@ build_ubuntu32()
 
 build_debian()
 {
+    if [ x"$1" = x'9' ]; then
+        true
+    else
+        QFSHADOOP_VERSIONS=$MYQFSHADOOP_VERSIONS_UBUNTU1804
+    fi
     build_ubuntu
 }
 
@@ -286,11 +293,18 @@ if [ $# -eq 5 -a x"$1" = x'build' ]; then
     exit
 fi
 
-if [ x"$TRAVIS_OS_NAME" = x'linux' ]; then
+if [ x"$TRAVIS_OS_NAME" = x ]; then
+    true
+else
+    if [ x"$BUILD_OS_NAME" = x ]; then
+        BUILD_OS_NAME=$TRAVIS_OS_NAME
+    fi
+fi
+
+if [ x"$BUILD_OS_NAME" = x'linux' ]; then
     if [ -e "$MYTMPDIR" ]; then
         rm -r "$MYTMPDIR"
     fi
-    make rat clean
     if [ x"$CODECOV" = x'yes' ]; then
         init_codecov
     fi
@@ -306,10 +320,14 @@ if [ x"$TRAVIS_OS_NAME" = x'linux' ]; then
     fi
     MYSRCD="$(pwd)"
     ulimit -c unlimited
-    docker run --rm --dns=8.8.8.8 -t -v "$MYSRCD:$MYSRCD" -w "$MYSRCD" \
-        "$DISTRO:$VER" \
-        /bin/bash ./travis/script.sh build "$DISTRO" "$VER" "$BTYPE" "$BUSER"
-elif [ x"$TRAVIS_OS_NAME" = x'osx' ]; then
+    if [ x"$BUILD_RUN_DOCKER" = x'no' ]; then
+        "$0" build "$DISTRO" "$VER" "$BTYPE" "$BUSER"
+    else
+        docker run --rm --dns=8.8.8.8 -t -v "$MYSRCD:$MYSRCD" -w "$MYSRCD" \
+            "$DISTRO:$VER" \
+            /bin/bash ./travis/script.sh build "$DISTRO" "$VER" "$BTYPE" "$BUSER"
+    fi
+elif [ x"$BUILD_OS_NAME" = x'osx' ]; then
     set_build_type "$BTYPE"
     for pkg_name in \
             'openssl@1.1' \
@@ -334,6 +352,6 @@ elif [ x"$TRAVIS_OS_NAME" = x'osx' ]; then
     df -h || true
     do_build -j 2
 else
-    echo "OS: $TRAVIS_OS_NAME not yet supported"
+    echo "OS: $BUILD_OS_NAME not yet supported"
     exit 1
 fi
